FBUtil for the login, and graph request/ issue on the setReadPermissions(public_profile) solved changing to setReadPermissions(email)
