+++ app/src/main/java/com/couchbase/todolite/util/FacebookResults.java
@@ -1,0 +1,10 @@
+package com.couchbase.todolite.util;
+
+import com.facebook.login.LoginResult;
+
+/**
+ * @author Moss
+ */
+public interface FacebookResults {
+    void onFbLoginSuccess(String token, String userId, String name);
+}
